+++ dsapy/src/algorithms/sorting/insertion_sort.py
@@ -1,64 +1,13 @@
-import time
+from dsapy.src.algorithms.sorting.sorting_algorithm import SortingAlgorithm
 
 
-class InsertionSort:
-    """
-    This class is an implementation of the Insertion Sort Algorithm.\n
-    It has the following methods:
-        - sort: The main implementation of the sort algorithm
-        - animation: Returns a Dictionary of Lists, containing all swap and comparison operations.\n
-        - metadata: Returns a Dictionary of metadata for the algorithm.
-
-    Animation Keys:
-        - swap
-        - comparison
-    """
+class InsertionSort(SortingAlgorithm):
 
     def __init__(self):
         """
-        Initializes the class with a metadata and an animation dictionary.
+        Initializes the base abstract class SortingAlgorithm
         """
-        self.__metadata = {}
-
-        self.__animation = {}
-
-    def __init_metadata(self):
-
-        self.__metadata = {
-            'swap_count': 0,
-            'comparison_count': 0,
-            'iteration_count': 0,
-            'runtime': 0.0,
-        }
-
-        self.__animation = {
-            'swap': [],
-            'comparison': []
-        }
-
-    def metadata(self) -> dict:
-        """
-        Returns a metadata dictionary.\n
-        Metadata Keys:
-            - swap_count
-            - comparison_count
-            - iteration_count
-            - run_time
-
-        :return: A metadata dictionary
-        """
-        return self.__metadata
-
-    def animation(self) -> dict:
-        """
-        Returns a dictionary containing all swaps and animations.\n
-        Animation Keys:
-            - swap
-            - comparison
-
-        :return: An animation dictionary
-        """
-        return self.__animation
+        super().__init__()
 
     def sort(self, data: list[int | float | str]) -> list[int | float | str]:
         """
@@ -71,39 +20,36 @@ class InsertionSort:
         """
 
         # Initiate/Reset the metadata and animations
-        self.__init_metadata()
+        self._reset_metadata_animations()
 
         # Copy the input to a new list
         data = data.copy()
 
         # CPU start time of the sorting algorithm
-        start_time = time.process_time()
+        self._start_runtime()
 
         # Loop through the list starting from the second element.
         for i in range(1, len(data)):
-            self.__metadata['iteration_count'] += 1
+            self._add_iteration()
             j = i
             # Loop through the sorted part of the list and place the new element in the correct place
             while j > 0:
-                self.__metadata['comparison_count'] += 1
-                self.__animation['comparison'].append([j, j + 1])
+                self._add_comparison()
                 # Transferred condition from while loop to make the comparisons animations
-                if data[j-1] <= data[j]:
+                if data[j - 1] <= data[j]:
                     break
 
                 # REAL while loop starts from this line
-                self.__metadata['iteration_count'] += 1
+                self._add_iteration()
 
                 # Swap
-                self.__metadata['swap_count'] += 1
-                self.__animation['swap'].append([j, j - 1])
+                self._add_swap(j, j-1)
                 data[j], data[j - 1] = data[j - 1], data[j]
                 j -= 1
 
         # CPU end time of the sorting algorithm
-        end_time = time.process_time()
-        # Total CPU time in ms
-        self.__metadata['runtime'] = round(1000 * (end_time - start_time), 2)
+        self._stop_runtime()
+        self._add_runtime()
 
         # Return the sorted array
         return data