+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [0.0.3] - 2022-12-01
 ### Changed
+- sorting_algorithm.py: replaced animation list with queue data structure
 - quick_sort.py: QuickSort Inherits the SortingAlgorithm abstract class
 - selection_sort.py: SelectionSort Inherits the SortingAlgorithm abstract class
 - insertion_sort.py: InsertionSort Inherits the SortingAlgorithm abstract class